+++ src/screens/authentication/signup/EmailSignUp.tsx
@@ -1,30 +1,26 @@
-import { View, Text, TouchableOpacity, Alert, ToastAndroid, Platform } from 'react-native'
+import { View, Text, TouchableOpacity, Alert, Platform, TextInput } from 'react-native'
 import React, { useState } from 'react'
-import { TextInput } from 'react-native-gesture-handler'
-import { useRoute } from '@react-navigation/native';
+import { useNavigation, useRoute } from '@react-navigation/native';
 import styles from '../../../utils/styles/shadow';
 import { supabase } from '../../../../supabase';
 import { useDispatch } from 'react-redux';
 import { setCurrentUser } from '../../../context/navSlice';
 import validateEmail from '../../../utils/functions/validateEmail';
-import { EmailSignUpScreenRouteProp } from '../../../utils/types/types';
+import { EmailSignUpScreenRouteProp, RootStackNavigationProp } from '../../../utils/types/types';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import validatePassword from '../../../utils/functions/validatePassword';
 
 const EmailSignUp = () => {
-	const route = useRoute<EmailSignUpScreenRouteProp>();
-	const { age, name } = route.params;
 	const [email, setEmail] = useState<string>('');
 	const [password, setPassword] = useState<string>('');
 	const [confirmPassword, setConfirmPassword] = useState<string>('');
 	const [loading, setLoading] = useState(false);
-	const dispatch = useDispatch();
+	const navigation = useNavigation<RootStackNavigationProp>();
 	const isPasswordMismatch = confirmPassword !== password;
 	const isEmailEmpty = email === '';
 	const isPasswordValid = validatePassword(password);
 
 	const isDisabled = isPasswordMismatch || isEmailEmpty || !isPasswordValid || loading;
-
 
 	async function signUpWithEmail() {
 		setLoading(true);
@@ -33,53 +29,38 @@ const EmailSignUp = () => {
 			setLoading(false);
 			return;
 		}
-		//Sign up the user to supabase
-		const {
-			data: { session },
-			error: AuthUserError,
-		} = await supabase.auth.signUp({
-			email: email,
-			password: password,
-		})
+		//Check if user already has an account
+		const { data: existingUser } = await supabase
+			.from('users')
+			.select('*')
+			.eq('email', email);
 
-		//If the sign up is successful, insert a row to public.users
-		if (session) {
-			await AsyncStorage.setItem('userAccessToken', session.access_token);
-			await AsyncStorage.setItem('userRefreshToken', session.refresh_token);
-			const { error, data } = await supabase
-				.from('users')
-				.insert({
-					name: name,
-					email: email,
-					uid: session?.user?.id,
-					age: age,
-					auth_provider: 'email'
-				})
-				.select('id')
-			if (error) { console.error(error.message); }
-			if (error?.code === '23505') {
-				Platform.OS === 'android' ?
-					ToastAndroid.show('You already have an account, just sign in!', ToastAndroid.SHORT)
-					:
-					Alert.alert('You already have an account, just sign in!')
-				return;
+		if (existingUser && existingUser.length > 0) {
+			alert("It looks like you already have an account. Please sign in instead.");
+			navigation.navigate('signin');
+			setLoading(false);
+		} else {
+			const {
+				data: { session },
+				error: AuthUserError,
+			} = await supabase.auth.signUp({
+				email: email,
+				password: password,
+			})
+			if ( AuthUserError) {
+				setLoading(false);
+				throw AuthUserError.message;
+			}
+			if (session) {
+				await AsyncStorage.setItem('userAccessToken', session.access_token);
+				await AsyncStorage.setItem('userRefreshToken', session.refresh_token);
+
+				navigation.navigate('userdetailsscreen');
+				if (!session) Alert.alert('Sorry, we could not sign you up');
+				setLoading(false);
 			}
 
-			//Once the new user has been added to public.users, add user properties to the context
-			if (data) {
-				dispatch(setCurrentUser({
-					name: name,
-					email: email,
-					// photo: userInfo.user.photo,
-					id: data[0].id
-				}))
-			}
 		}
-
-
-		if (AuthUserError) Alert.alert(AuthUserError.message);
-		if (!session) Alert.alert('Sorry, we could not sign you up');
-		setLoading(false);
 	}
 
 	return (
@@ -103,7 +84,6 @@ const EmailSignUp = () => {
 					value={email}
 					onChangeText={(value) => { setEmail(value) }}
 				/>
-
 			</View>
 			<View className='w-5/6 space-y-1'>
 				<Text className='ml-2 text-lg font-bold'>
@@ -116,7 +96,6 @@ const EmailSignUp = () => {
 					className={`p-2 px-5 flex items-center border  rounded-full bg-gray-200 ${Platform.OS === 'ios' ? 'py-4' : 'py-2'}`}
 					onChangeText={(value) => { setPassword(value) }}
 				/>
-
 			</View>
 			<View className='w-5/6 space-y-1'>
 				<Text className='ml-2 text-lg font-bold'>
@@ -131,7 +110,7 @@ const EmailSignUp = () => {
 				/>
 
 			</View>
-			{(!isPasswordValid && password !== '')&&  (
+			{(!isPasswordValid && password !== '') && (
 				<Text className='text-red-500 w-5/6 text-center'>
 					Password must be 6-16 characters long, contain at least one number and one special character.
 				</Text>
